move reminder email job into email_jobs
