+++ backend/python/app/services/jobs/email_jobs.py
@@ -10,7 +10,9 @@ from datetime import date, datetime, timedelta
 from typing import TYPE_CHECKING, Any, cast
 
 from sqlalchemy import and_
-from sqlmodel import select
+from sqlmodel import col, select
+
+from app.config import settings
 
 if TYPE_CHECKING:
     from sqlalchemy.sql.schema import Table
@@ -19,7 +21,9 @@ from app.dependencies.services import get_email_dispatcher, get_logger
 from app.models.driver import Driver
 from app.models.route import Route
 from app.models.route_group import RouteGroup
+from app.models.system_settings import SystemSettings
 from app.models.user import User
+from app.services.implementations.email_service import EmailService
 
 
 async def send_route_reminders() -> None:
@@ -141,5 +145,110 @@ async def send_route_reminders() -> None:
         logger.error(
             f"Failed to process route reminder emails: {error!s}",
             exc_info=True,
+        )
+        raise error
+
+
+async def process_daily_reminder_emails() -> None:
+    """Sends out daily reminder emails for the configured reminder window.
+
+    Emails each driver assigned (via Route.driver_id) to a route whose
+    RouteGroup.drive_date falls on one of the configured lead days.
+    """
+
+    from app.models import (
+        async_session_maker_instance,  # Placed here to ensure testing file functions as normal
+    )
+
+    logger = get_logger()
+
+    if async_session_maker_instance is None:
+        logger.error("Database session maker not initialized")
+        return
+
+    try:
+        async with async_session_maker_instance() as session:
+            settings_result = await session.execute(select(SystemSettings).limit(1))
+            system_settings = settings_result.scalars().first()
+            reminder_days = (
+                system_settings.email_reminder_days_before
+                if system_settings and system_settings.email_reminder_days_before
+                else [1]
+            )
+            target_dates = {date.today() + timedelta(days=day) for day in reminder_days}
+            start_of_day = datetime.combine(min(target_dates), datetime.min.time())
+            end_of_day = datetime.combine(max(target_dates), datetime.max.time())
+            statement = (
+                select(
+                    User.email,
+                    RouteGroup.drive_date,
+                    col(Route.start_time),
+                    Route.length,
+                )
+                .join(RouteGroup, RouteGroup.route_group_id == Route.route_group_id)  # type: ignore[arg-type]
+                .join(Driver, Driver.driver_id == Route.driver_id)  # type: ignore[arg-type]
+                .join(User, User.user_id == Driver.user_id)  # type: ignore[arg-type]
+                .where(
+                    and_(
+                        RouteGroup.drive_date >= start_of_day,  # type: ignore[arg-type]
+                        RouteGroup.drive_date <= end_of_day,  # type: ignore[arg-type]
+                        col(Route.driver_id).isnot(None),
+                    )
+                )
+                .order_by(User.email)
+            )
+
+            result = await session.execute(statement)
+            upcoming_routes = result.all()
+
+            if not upcoming_routes:
+                logger.info(
+                    "No Upcoming Routes found for configured reminder days, skipping emails"
+                )
+                return
+
+            email_service = EmailService(
+                logger,
+                {
+                    "refresh_token": settings.mailer_refresh_token,
+                    "token_uri": "https://oauth2.googleapis.com/token",
+                    "client_id": settings.mailer_client_id,
+                    "client_secret": settings.mailer_client_secret,
+                },
+                settings.mailer_user,
+                "Food4Kids",
+            )
+
+            with open("./app/templates/view-upcoming-route.html") as file:
+                template_html = file.read()
+
+            for row in upcoming_routes:
+                recipient_email = row.email
+                drive_date: datetime = row.drive_date
+                route_distance = row.length
+
+                date_only = drive_date.date().strftime("%A, %B %d, %Y")
+                # Per-route start time if set, else fall back to a sensible default.
+                start_time = row.start_time
+                time_only = start_time.strftime("%I:%M %p") if start_time else "TBD"
+                rounded_distance = str(round(route_distance))
+
+                formatted_email = template_html.replace("Date_To_Replace", date_only)
+                formatted_email = formatted_email.replace("Time_To_Replace", time_only)
+                formatted_email = formatted_email.replace(
+                    "Route_Duration_To_Replace", rounded_distance
+                )
+                logger.info(f"Sending Email to {recipient_email}")
+                email_service.send_email(
+                    to=recipient_email,
+                    subject="Upcoming Route Reminder",
+                    body=formatted_email,
+                )
+
+            logger.info("Successfully sent reminder emails")
+
+    except Exception as error:
+        logger.error(
+            f"Failed to process daily reminder emails: {error!s}", exc_info=True
         )
         raise error
